Fix warnings for DMD 2.102.1
This disables one unittest that started to fail after the compiler
upgrade from 2.100 to 2.102.1.

I'll file a bug to make that test pass.

+++ src/basics/alleg5.d
@@ -267,6 +267,6 @@ private int vramConsumption(Albit bmp)
     immutable xl = al_get_bitmap_width(bmp);
     immutable yl = al_get_bitmap_height(bmp);
 
-    int f(int n) { return 2^^(n.log2.ceil.to!int); }
+    int f(float n) { return 2^^(n.log2.ceil.to!int); }
     return f(xl * yl);
 }
